% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaPlot.R
\name{areaPlot}
\alias{areaPlot}
\title{areaPlot}
\usage{
areaPlot(
  data,
  title = NULL,
  caption = FALSE,
  grid = FALSE,
  ticks = TRUE,
  sp500 = FALSE,
  font
)
}
\arguments{
\item{data}{Data frame/data table/tibble with each year's datat point as a row and columns containing year and variable to visualize.}

\item{title}{Naming the chart (e.g. "Unfunded Liability Growth").}

\item{caption}{Set to TRUE to add "reason.org/pension" caption at the bottom right corner}

\item{grid}{Set to TRUE to add major gridlines}

\item{ticks}{Set to FALSE to remove ticks}

\item{sp500}{Default (FALSE) visualizes pension debt. Set to TRUE to visualize annnual S&P500 Index values on the secondary Y-axis}

\item{font}{Directly paste name of a font (e.g. "Calibri") to change the default font of the text}
}
\description{
areaPlot
}
\examples{
\dontrun{
areaPlot(data, caption = FALSE, grid = FALSE, ticks = TRUE, sp500 = FALSE, font = "Calibri")
}
}
\author{
Anil Niraula <anil.niraula@reason.org>
}
