% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAssets.R
\name{assetPlot}
\alias{assetPlot}
\title{Plot Asset data}
\usage{
assetPlot(data, interactive = FALSE)
}
\arguments{
\item{data}{a data frame or data.table containing PPD's asset allocation data.}

\item{interactive}{set to `TRUE` to create interactive version of the chart.}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
Plot Asset data
}
\examples{
\dontrun{
assetPlot(data = assets, interactive = FALSE)
}
}
\author{
Anil Niraula <anil.niraula@reason.org>
}
