% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_plot.R
\name{debtPlot}
\alias{debtPlot}
\title{Create the 'Mountain of Debt' plot.}
\usage{
debtPlot(data)
}
\arguments{
\item{data}{a dataframe produced by the selectedData function or in the same format containing year, uaal, funded ratio columns.}
}
\description{
Create the 'Mountain of Debt' plot.
}
\examples{
\dontrun{
debtPlot(data)
}
}
