% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_plot.R
\name{debtTable}
\alias{debtTable}
\title{Create a table comntaining the data used in the 'Mountain of Debt' plot.}
\usage{
debtTable(data)
}
\arguments{
\item{data}{a dataframe produced by the selectedData function or in the same format containing year, uaal, funded ratio columns.}
}
\description{
Create a table comntaining the data used in the 'Mountain of Debt' plot.
}
\examples{
\dontrun{
debtTable(data)
}
}
