% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dw}
\alias{dw}
\title{Connection credientials.}
\format{A list with 9 items.}
\usage{
dw
}
\description{
A dataset containing the connection parameters in a list format to be used in connecting to the Reason pension database.
}
\keyword{datasets}
