% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filter data
devtools::document()}
\usage{
filter(data, source = FALSE)
}
\arguments{
\item{data}{data file (in data.frame/data.table format)}

\item{source}{TRUE/FALSE statement to indicate if "source data" is used}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
Filter data
devtools::document()
}
\examples{
\dontrun{
filter(data, source = FALSE)
}
}
