% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAssets.R
\name{filterAssets}
\alias{filterAssets}
\title{Filter Asset data}
\usage{
filterAssets(data, fy = 2001, plan = NULL)
}
\arguments{
\item{data}{a data frame or data.table containing PPD's asset allocation data.}

\item{fy}{Numeric parameter containing a starting fiscal year as it is listed in the Reason pension database.}

\item{plan}{name of state-managed pension plan to filter end data for.}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
Filter Asset data
}
\examples{
\dontrun{
filterAssets(data, fy = 2001, plan = NULL)
}
}
\author{
Anil Niraula <anil.niraula@reason.org>
}
