% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{filterData
devtools::document()}
\usage{
filterData(Data, fy, source = FALSE)
}
\arguments{
\item{source}{TRUE/FALSE statement to indicate if "source data" is used}

\item{data}{data file (in data.frame/data.table format)}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
filterData
devtools::document()
}
\examples{
\dontrun{
filter(data, 2001, source = FALSE)
}
}
