% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredData.R
\name{filteredData}
\alias{filteredData}
\title{filteredData filters around 45 columns, starting year, adds missing columns, and populates DR column}
\usage{
filteredData(Data, fy, source = FALSE)
}
\arguments{
\item{Data}{data.table or data.frame with pension values (see pullData()/pullStateData())}

\item{fy}{starting fiscal year}

\item{source}{TRUE if Data coemes from using pullSourceData() function}
}
\value{

}
\description{
filteredData filters around 45 columns, starting year, adds missing columns, and populates DR column
}
\examples{
\dontrun{
filteredData(Data, 2001, FALSE)
}
}
