% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainloss_plot.R
\name{glPlot}
\alias{glPlot}
\title{Create the 'Gain/Loss' plot using a file as an input.}
\usage{
glPlot(filename, ylab_unit = "Billions")
}
\arguments{
\item{filename}{a csv (comma separated value) file containing columns of gain loss category names with one row of values.}

\item{ylab_unit}{a string contained within quotation marks containing th y-axis label unit. The default value is "Billions"}
}
\description{
Create the 'Gain/Loss' plot using a file as an input.
}
\examples{
\dontrun{
glPlot(filename = 'data/Graph 1.csv')
glPlot(filename = 'data/Graph 2.csv', ylab = "Millions")
glPlot(filename = 'data/Graph 2.csv', ylab = "millions")
}
}
