% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{linePlot}
\alias{linePlot}
\title{Create a plot comparing two variables, such as ADEC vs. Actual contributions.}
\usage{
linePlot(data, .var1 = "adec_contribution_rates",
  .var2 = "actual_contribution_rates",
  labelY = "Employer Contribution (\% of Payroll)",
  label1 = "ADEC Contribution Rate",
  label2 = "Actual Contribution Rate")
}
\arguments{
\item{data}{a dataframe produced by the selectedData function or in the same format.}

\item{.var1}{The name of the first variable to plat, default is adec_contribution_rates.}

\item{.var2}{The name of the second variable to plot, default if actual_contribution_rates.}

\item{labelY}{A label for the Y-axis.}

\item{label1}{A label for the first variable.}

\item{label2}{A label for the second variable.}
}
\description{
Create a plot comparing two variables, such as ADEC vs. Actual contributions.
}
