% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{loadData}
\alias{loadData}
\title{Load the data for a specified plan from an Excel file.}
\usage{
loadData(filename)
}
\arguments{
\item{filename}{A string enclosed in quotation marks containing a file name with path of a pension plan Excel data file.}
}
\description{
Load the data for a specified plan from an Excel file.
}
\examples{
\dontrun{
allWide <- loadData('data/NorthCarolina_PensionDatabase_TSERS.xlsx')
}
}
