% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterView.R
\name{masterView}
\alias{masterView}
\title{View mapped/unmapped columns per data source}
\usage{
masterView(source = NULL, expand = FALSE)
}
\arguments{
\item{source}{data source (e.g. "Reason", "Public Plans Database", "Census").}

\item{expand}{TRUE or FALSE (default) allowing to see either currently mapped or all unmapped variables}
}
\value{
A wide data frame with each column names, sources, and other details.
}
\description{
View mapped/unmapped columns per data source
}
\examples{
\dontrun{
masterView("Reason", TRUE)
}
}
\author{
Anil Niraula <anil.niraula@reason.org>
}
