% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{modData}
\alias{modData}
\title{Select the data used in the 'mountain of debt' graph.}
\usage{
modData(wide_data, .year_var = "year",
  .aal_var = "actuarial_accrued_liabilities_under_gasb_standards",
  .asset_var = "actuarial_assets_under_gasb_standards", base = 1000)
}
\arguments{
\item{wide_data}{a datasource in wide format}

\item{.year_var}{the name of the column conatining the year}

\item{.aal_var}{the name of the column containing the AAL, default is Reason db column name}

\item{.asset_var}{the name of the column containing the Actuarial Assets, default to Reason db name.}

\item{base}{Does the plan report their numbers by the thousand dollar or by the dollar? default is 1000, change to 1 for plans that report by the dollar}
}
\value{
A data frame containing the columns: year, actuarial_assets, aal, uaal, funded_ratio
}
\description{
Select the data used in the 'mountain of debt' graph.
}
\examples{
\dontrun{
data <- modData(wide_data,
                .year_var = 'Fiscal Year End',
                .aal_var = 'Actuarial Accrued Liability',
                .asset_var = 'Actuarial Value of Assets',
                 base = 1
                 )
}
}
