% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.colors.R
\docType{data}
\name{palette_reason}
\alias{palette_reason}
\title{Reason palette.}
\format{
A data frame with 14+ variables/items.
}
\usage{
palette_reason
}
\description{
A data.frame containing the rgb codes to reference Reason colors in the code.
}
\keyword{datasets}
