% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensionviewr.R
\docType{package}
\name{pensionviewr}
\alias{pensionviewr}
\alias{pensionviewr-package}
\title{pensionviewr: Pension Data Pull/Ananlysis (Reason Foundation)}
\description{
Allows to pull and filter data from the pension database
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/ReasonFoundation/pensionviewr}
  \item Report bugs at \url{http://github.com/ReasonFoundation/pensionviewr/issues}
}

}
\author{
\strong{Maintainer}: Anil Niraula \email{anil.niraula@reason.org}

Authors:
\itemize{
  \item Andrew Abbott \email{andrew.abbott@reason.org}
}

}
