% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{planList}
\alias{planList}
\title{Grab a list of the plans with their state from the Reason database.}
\usage{
planList()
}
\value{
A dataframe containing the pension plan id, name, and state.
}
\description{
Grab a list of the plans with their state from the Reason database.
}
\examples{
planList()
}
