% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullAssetData.R
\name{pullAssetData}
\alias{pullAssetData}
\title{pullAssetData}
\usage{
pullAssetData(fy, mva = FALSE)
}
\arguments{
\item{fy}{starting fiscal year to filter data for}

\item{mva}{= FALSE (Set to `TRUE` to also pull MVA for weighting)}
}
\description{
pullAssetData
}
\examples{
\dontrun{
pullAssetData(fy = 2001, mva = FALSE) 
}
}
\author{
Anil Niraula <anil.niraula@reason.org>
}
