% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{pullData}
\alias{pullData}
\title{Pull the data for a specified plan from the Reason pension databse.}
\usage{
pullData(pl, plan_name = "Texas Employees Retirement System")
}
\arguments{
\item{pl}{A datafram containing the list of plan names, states, and ids in the form produced by the planList() function.}

\item{plan_name}{A string enclosed in quotation marks containing a plan name as it is listed in the Reason pension database.}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
Pull the data for a specified plan from the Reason pension databse.
}
\examples{
\dontrun{
pullData(pl)
pullData(pl, "Kansas Public Employees' Retirement System")
}
}
