% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullSourceData.R
\name{pullSourceData}
\alias{pullSourceData}
\title{Pull data sources}
\usage{
pullSourceData(plan_name)
}
\arguments{
\item{plan_name}{}
}
\value{

}
\description{
Pulls each data point with all data sources this data point has
}
\examples{
pullSourceData(pl, "New Mexico Educational Retirement Board")
}
