% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullStateData.R
\name{pullStateData}
\alias{pullStateData}
\title{Pull all data for state-level pension plans}
\usage{
pullStateData(state, FY)
}
\arguments{
\item{state}{}

\item{FY}{}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
Pull all data for state-level pension plans
}
\examples{
\dontrun{
pullStateData(state, 2001)
}
}
