% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reasonTheme}
\alias{reasonTheme}
\title{A Reason theme for ggplot2.}
\format{A list with 10 items.}
\usage{
reasonTheme
}
\description{
A dataset containing a Reason theme in a list format to be used in creating ggplot2 plots.
}
\keyword{datasets}
