% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{savePlot}
\alias{savePlot}
\title{Add a source and save ggplot chart}
\usage{
savePlot(plot, source, save_filepath = file.path(Sys.getenv("TMPDIR"),
  "tmp-nc.png"), width_pixels = 648, height_pixels = 384.48)
}
\arguments{
\item{plot}{The variable name of the plot you have created that you want to format and save}

\item{source}{The text you want to come after the text 'Source:' in the bottom left hand side of your side}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{width_pixels}{Width in pixels that you want to save your chart to - defaults to 648}

\item{height_pixels}{Height in pixels that you want to save your chart to - defaults to 384.48
which needs to be a PNG file - defaults to BBC blocks image that sits within the data folder of your package}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Running this function will save your plot with the correct guidelines for publication.
It will left align your source and save it to your specified location.
}
\examples{
\dontrun{
savePlot(plot = myplot,
source = "The source for my data",
save_filepath = "filename_that_my_plot_should_be_saved_to.png",
width_pixels = 648,
height_pixels = 384.48)
}

}
