% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{selectedData}
\alias{selectedData}
\title{Selects the data used in several graphs.}
\usage{
selectedData(wide_data)
}
\arguments{
\item{wide_data}{a datasource in wide format}
}
\value{
A data frame containing the columns: year, valuation_date, actuarial_assets, aal, adec, er_cont, ee_cont, payroll, uaal, funded_ratio, adec_contribution_rates, actual_contribution_rates
}
\description{
Selects the data used in several graphs.
}
\examples{
\dontrun{
data <- selected_Data(wide_data)
}
}
