% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_plot.R
\name{contPlot}
\alias{contPlot}
\title{Create a plot comparing ADEC contribution rates with actual contribution rates.}
\usage{
contPlot(data, y1 = "ADEC Contribution Rates",
  y2 = "Actual Contribution Rates (Statutory)", y3 = NULL,
  labelY = NULL, label1 = NULL, label2 = NULL, label3 = NULL)
}
\arguments{
\item{data}{a dataframe produced by the selectedData function or in the same format containing columns for ADEC, actual and optionally a third column to plot.}

\item{y1}{The name of the ADEC variable}

\item{y2}{The name of the actual contribution variable}

\item{y3}{The name of an optional third variable}

\item{labelY}{A label for the Y-axis}

\item{label1}{A label for the ADEC variable}

\item{label2}{A label for the actual contribution variable}

\item{label3}{A label for the third variable}
}
\description{
Create a plot comparing ADEC contribution rates with actual contribution rates.
}
