% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_plot.R
\name{contPlot}
\alias{contPlot}
\title{Create a plot comparing ADEC contribution rates with actual contribution rates.}
\usage{
contPlot(data, .adec_var = "adec_contribution_rates",
  .actual_var = "actual_contribution_rates", labelY = NULL,
  label1 = "ADEC Contribution Rate",
  label2 = "Actual Contribution Rate")
}
\arguments{
\item{data}{a dataframe produced by the selectedData function or in the same format containing columns for ADEC, and actual contributions to plot.}

\item{labelY}{A label for the Y-axis}

\item{label1}{A label for the ADEC variable}

\item{label2}{A label for the actual contribution variable}

\item{y1}{The name of the ADEC variable}

\item{y2}{The name of the actual contribution variable}
}
\description{
Create a plot comparing ADEC contribution rates with actual contribution rates.
}
