% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{FilterData filters data
devtools::document()}
\usage{
filterData(data, fy, source = FALSE)
}
\arguments{
\item{data}{data file (in data.frame/data.table format)}

\item{fy}{starting fiscal year}

\item{source}{TRUE/FALSE statement to indicate if "source data" is used}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
FilterData filters data
devtools::document()
}
\examples{
\dontrun{
filter(data, 2001, source = FALSE)
}
}
