% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{filterData selects around 50 of the original columns, recreates any of the missing columns, 
renames column headers for simplicity, and allows to filter plans by type of employees covered}
\usage{
filterData(Data, fy = 2001, employee = NULL, source = FALSE)
}
\arguments{
\item{Data}{data file (in data.frame/data.table format)}

\item{fy}{starting year}

\item{employee}{character parameter designating filter for 
type of employees covered (e.g. "teacher", "state", "local", "state and local", "police and fire").}

\item{source}{TRUE/FALSE statement to indicate if "source data" is used}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
filterData selects around 50 of the original columns, recreates any of the missing columns, 
renames column headers for simplicity, and allows to filter plans by type of employees covered
}
\examples{
\dontrun{
filterData(Data, 2001, employee = "teacher", source = FALSE)
}
}
\author{
Anil Niraula <anil.niraula@reason.org>
}
