% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linePlot.R
\name{linePlot}
\alias{linePlot}
\title{Create a line plot to visualize up to 5 variables, such as ADEC vs. Actual contributions.}
\usage{
linePlot(
  data,
  title = NULL,
  caption = FALSE,
  grid = FALSE,
  interactive = FALSE,
  treasury = FALSE,
  inv.returns = TRUE,
  ticks = TRUE,
  font = NULL,
  yaxisMin = 0,
  yaxisMax = NULL,
  yaxisSeq = 5,
  yaxisScale = 100,
  Ytitle = NULL,
  format = NULL,
  str = 20,
  labelY = NULL,
  lab1 = NULL,
  lab2 = NULL,
  lab3 = NULL,
  lab4 = NULL,
  lab5 = NULL
)
}
\arguments{
\item{data}{a dataframe/data.table produced by pullData, pullStateData, or other ways that produces the same format.}

\item{title}{Naming the chart (e.g. "Unfunded Liability Growth").}

\item{caption}{Set to TRUE to add "reason.org/pension" caption at the bottom right corner}

\item{grid}{Set to TRUE to add major gridlines}

\item{interactive}{Set to TRUE to create interactive version of the plot (w/ plotly)}

\item{treasury}{= Set to TRUE to show 30-Year treasury Yeilds on the secondary Y-axis}

\item{inv.returns}{= Set to FALSE to graph anything other than Investsment Return graph}

\item{ticks}{Set to FALSE to remove ticks#'}

\item{font}{Directly paste name of a font (e.g. "Calibri") to change the default font of the text}

\item{yaxisMin}{Value that sets Y-axis minimum.}

\item{yaxisMax}{Value that sets Y-axis maximum.}

\item{yaxisSeq}{Value that sets space between Major breaks.}

\item{yaxisScale}{Value that sets Y-axis scale. Example: 100 for percentages or 1/1000 for thousands.}

\item{Ytitle}{Set Y-axis title}

\item{format}{Format of Y-axis scale. Examples: "%", "$", or something else.}

\item{str}{Value that sets number of strings at which to cut legend text at -- Default is 20.}

\item{labelY}{Title of the Y-axis.}

\item{lab1}{Text label for the 1st variable - optional.}

\item{lab2}{Text label for the 2nd variable - optional.}

\item{lab3}{Text label for the 3rd variable - optional.}

\item{lab4}{Text label for the 4th variable - optional.}

\item{lab5}{Text label for the 5th variable - optional.}
}
\description{
Create a line plot to visualize up to 5 variables, such as ADEC vs. Actual contributions.
}
\examples{
\dontrun{
graph <- linePlot(ERB.data, title = "<b>A History of New Mexico Investment Returns (2001-2020)<b>", inv.returns = F, treasury = T,
font = "Arial", yaxisMin = -21, yaxisMax = 21, yaxisSeq = 3,
interactive = T, Ytitle = NULL,
yaxisScale = 100, format = "\%", str = 60,
labelY = "", lab1 = "Market Valued Returns (Actual)", 
lab2 = "Actuarially Valued Investment Return (Smoothed by Plan)", 
lab3 = "Assumed Rate of Return")
}
}
\author{
Anil Niraula <anil.niraula@reason.org>, Swaroop Bhagavatula <swaroop.bhagavatula@reason.org>, Jen Sidorova <jen.sidorova@reason.org>
}
