% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linePlot.R
\name{linePlot}
\alias{linePlot}
\title{Create a line plot to visualize up to 5 variables, such as ADEC vs. Actual contributions.}
\usage{
linePlot(
  data,
  Min = 0,
  Max = NULL,
  Seq = 5,
  Scale = 100,
  format = NULL,
  str = 20,
  labelY = NULL,
  lab1 = NULL,
  lab2 = NULL,
  lab3 = NULL,
  lab4 = NULL,
  lab5 = NULL
)
}
\arguments{
\item{data}{a dataframe/data.table produced by pullData, pullStateData, or other ways that produces the same format.}

\item{Min}{Value that sets Y-axis minimum.}

\item{Max}{Value that sets Y-axis maximum.}

\item{Seq}{Value that sets space between Major breaks.}

\item{Scale}{Value that sets Y-axis scale. Example: 100 for percentages or 1/1000 for thousands.}

\item{format}{Format of Y-axis scale -- either "%", "$", or something else.}

\item{str}{Value that sets number of strings at which to cut legend text at. Default is 20.}

\item{labelY}{Title of the Y-axis.}

\item{lab1}{Text label for the 1st variable -- optional.}

\item{lab2}{Text label for the 2nd variable -- optional.}

\item{lab3}{Text label for the 3rd variable -- optional.}

\item{lab4}{Text label for the 4th variable -- optional.}

\item{lab5}{Text label for the 5th variable -- optional.}
}
\description{
Create a line plot to visualize up to 5 variables, such as ADEC vs. Actual contributions.
}
\author{
Anil Niraula <anil.niraula@reason.org>

Swaroop Bhagavatula <swaroop.bhagavatula@reason.org>

Jen Sidorova <jen.sidorova@reason.org>
}
