% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullSourceData.R
\name{pullSourceData}
\alias{pullSourceData}
\title{Pull source data}
\usage{
pullSourceData(pl, plan_name)
}
\arguments{
\item{pl}{A datafram containing the list of plan names, states, and ids in the form produced by the planList() function.}

\item{plan_name}{A string enclosed in quotation marks containing a plan name as it is listed in the Reason pension database.}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
Pull source data
}
\examples{
\dontrun{
pullSourceData(pl)
pullSourceData(pl, "Kansas Public Employees' Retirement System")
}
}
