% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullStateData.R
\name{pullStateData}
\alias{pullStateData}
\title{Pull State data}
\usage{
pullStateData(FY)
}
\arguments{
\item{FY}{Numeric parameter containing a starting fiscal year as it is listed in the Reason pension database.}
}
\value{
A wide data frame with each year as a row and variables as columns.
}
\description{
Pull State data
}
\examples{
\dontrun{
pullStateData(2001)
}
}
\author{
Anil Niraula <anil.niraula@reason.org>
}
